Fix conflicts

+++ src/main/java/com/restaurant/commands/request/ReservationDTO.java
@@ -11,6 +11,6 @@ public class ReservationDTO {
 
     private LocalDateTime dateReservation;
     private String comments;
-    private Integer tableId;
+    private Long tableId;
     private Long orderId;
 }
