+++ src/main/java/com/restaurant/views/ReservationView.java
@@ -12,7 +12,7 @@ public class ReservationView {
     private Long reservationId;
     private LocalDateTime dateReservation;
     private String comments;
-    private Integer tableId;
+    private Long tableId;
     private Long orderId;
 
 }
